+++ vmware_horizon.py
@@ -2465,7 +2465,7 @@ class Entitlements:
         """Lists the user or group entitlements for Global Desktop Entitlements in the environment.
 
         For information on filtering see https://vdc-download.vmware.com/vmwb-repository/dcr-public/f92cce4b-9762-4ed0-acbd-f1d0591bd739/235dc19c-dabd-43f2-8d38-8a7a333e914e/HorizonServerRESTPaginationAndFilterGuide.doc
-        Available for Horizon 8 2006 and later."""
+        Available for Horizon 8 2012 and later."""
 
         def int_get_global_desktop_entitlements(self, page:int, maxpagesize: int, filter:list="") ->list:
             if filter != "":
@@ -2499,11 +2499,11 @@ class Entitlements:
             results += response.json()
         return results
 
-    def add_global_desktop_entitlement(self, global_desktop_entitlement_data: dict):
+    def new_global_desktop_entitlement(self, global_desktop_entitlement_data: list):
         """Create the bulk entitlements for a set of Global Desktop Entitlements.
 
         Requires global_desktop_entitlement_data as a dict
-        Available for Horizon 8 2006 and later."""
+        Available for Horizon 8 2012 and later."""
         headers = self.access_token
         headers["Content-Type"] = 'application/json'
         json_data = json.dumps(global_desktop_entitlement_data)
@@ -2513,7 +2513,7 @@ class Entitlements:
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
             raise Exception(f"Error {response.status_code}: {response.reason}")
-        elif response.status_code != 201:
+        elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
             try:
@@ -2521,11 +2521,11 @@ class Entitlements:
             except requests.exceptions.RequestException as e:
                 raise "Error: " + str(e)
 
-    def delete_global_desktop_entitlement(self, global_desktop_entitlement_data: dict):
+    def delete_global_desktop_entitlement(self, global_desktop_entitlement_data: list):
         """Delete the bulk entitlements for a set of Global Desktop Entitlements.
 
         Requires global_desktop_entitlement_data as a dict
-        Available for Horizon 8 2006 and later."""
+        Available for Horizon 8 2012 and later."""
         headers = self.access_token
         headers["Content-Type"] = 'application/json'
         json_data = json.dumps(global_desktop_entitlement_data)
@@ -2535,10 +2535,228 @@ class Entitlements:
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
             raise Exception(f"Error {response.status_code}: {response.reason}")
-        elif response.status_code != 201:
+        elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
             try:
                 response.raise_for_status()
             except requests.exceptions.RequestException as e:
                 raise "Error: " + str(e)
+
+    def get_global_application_entitlement(self, global_application_entitlement_id:str) -> dict:
+        """Gets the user or group entitlements for a Global Application Entitlement.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/entitlements/v1/global-application-entitlements/{global_application_entitlement_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def get_global_application_entitlements(self, maxpagesize:int=100, filter:dict="") -> list:
+        """Lists the user or group entitlements for Global Application Entitlements in the environment.
+
+        For information on filtering see https://vdc-download.vmware.com/vmwb-repository/dcr-public/f92cce4b-9762-4ed0-acbd-f1d0591bd739/235dc19c-dabd-43f2-8d38-8a7a333e914e/HorizonServerRESTPaginationAndFilterGuide.doc
+        Available for Horizon 8 2012 and later."""
+
+        def int_get_global_application_entitlements(self, page:int, maxpagesize: int, filter:list="") ->list:
+            if filter != "":
+                add_filter = urllib.parse.quote(json.dumps(filter,separators=(', ', ':')))
+                response = requests.get(f'{self.url}/rest/entitlements/v1/global-application-entitlements?filter={add_filter}&page={page}&size={maxpagesize}', verify=False, headers=self.access_token)
+            else:
+                response = requests.get(f'{self.url}/rest/entitlements/v1/global-application-entitlements?page={page}&size={maxpagesize}', verify=False, headers=self.access_token)
+            if response.status_code == 400:
+                if "error_messages" in response.json():
+                    error_message = (response.json())["error_messages"]
+                else:
+                    error_message = (response.json())["error_message"]
+                raise Exception(f"Error {response.status_code}: {error_message}")
+            elif response.status_code != 200:
+                raise Exception(f"Error {response.status_code}: {response.reason}")
+            else:
+                try:
+                    response.raise_for_status()
+                except requests.exceptions.RequestException as e:
+                    raise "Error: " + str(e)
+                else:
+                    return response
+        if maxpagesize > 1000:
+            maxpagesize = 1000
+        page = 1
+        response = int_get_global_application_entitlements(self,page = page, maxpagesize= maxpagesize,filter = filter)
+        results = response.json()
+        while 'HAS_MORE_RECORDS' in response.headers:
+            page += 1
+            response = int_get_global_application_entitlements(self,page = page, maxpagesize= maxpagesize, filter = filter)
+            results += response.json()
+        return results
+
+    def new_global_application_entitlement(self, global_application_entitlement_data: list):
+        """Create the bulk entitlements for a set of Global Application Entitlements
+
+        Requires global_application_entitlement_data as a dict
+        Available for Horizon 8 2012 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        json_data = json.dumps(global_application_entitlement_data)
+        response = requests.post(f'{self.url}/rest/entitlements/v1/global-application-entitlements', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+
+    def delete_global_application_entitlement(self, global_application_entitlement_data: list):
+        """Delete the bulk entitlements for a set of Global Application Entitlements
+
+        Requires global_application_entitlement_data as a dict
+        Available for Horizon 8 2012 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        json_data = json.dumps(global_application_entitlement_data)
+        response = requests.delete(f'{self.url}/rest/entitlements/v1/global-application-entitlements', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+
+class Federation:
+    def __init__(self, url: str, access_token: dict):
+        """Default object for the pools class where all Desktop Pool Actions will be performed."""
+        self.url = url
+        self.access_token = access_token
+
+    def get_cloud_pod_federation(self) -> dict:
+        """Retrieves the pod federation details.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/federation/v1/cpa', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def get_home_sites(self, maxpagesize:int=100, filter:dict="") -> list:
+        """Lists all the home sites in the pod federation.
+
+        For information on filtering see https://vdc-download.vmware.com/vmwb-repository/dcr-public/f92cce4b-9762-4ed0-acbd-f1d0591bd739/235dc19c-dabd-43f2-8d38-8a7a333e914e/HorizonServerRESTPaginationAndFilterGuide.doc
+        Available for Horizon 8 2012 and later."""
+
+        def int_Get_home_sites(self, page:int, maxpagesize: int, filter:list="") ->list:
+            if filter != "":
+                add_filter = urllib.parse.quote(json.dumps(filter,separators=(', ', ':')))
+                response = requests.get(f'{self.url}/rest/federation/v1/home-sites?filter={add_filter}&page={page}&size={maxpagesize}', verify=False, headers=self.access_token)
+            else:
+                response = requests.get(f'{self.url}/rest/federation/v1/home-sites?page={page}&size={maxpagesize}', verify=False, headers=self.access_token)
+            if response.status_code == 400:
+                if "error_messages" in response.json():
+                    error_message = (response.json())["error_messages"]
+                else:
+                    error_message = (response.json())["error_message"]
+                raise Exception(f"Error {response.status_code}: {error_message}")
+            elif response.status_code != 200:
+                raise Exception(f"Error {response.status_code}: {response.reason}")
+            else:
+                try:
+                    response.raise_for_status()
+                except requests.exceptions.RequestException as e:
+                    raise "Error: " + str(e)
+                else:
+                    return response
+        if maxpagesize > 1000:
+            maxpagesize = 1000
+        page = 1
+        response = int_Get_home_sites(self,page = page, maxpagesize= maxpagesize,filter = filter)
+        results = response.json()
+        while 'HAS_MORE_RECORDS' in response.headers:
+            page += 1
+            response = int_Get_home_sites(self,page = page, maxpagesize= maxpagesize, filter = filter)
+            results += response.json()
+        return results
+
+    def get_sites(self) -> list:
+        """Lists all the sites in the pod federation.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/federation/v1/sites', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def get_site(self, site_id:str) -> dict:
+        """Retrives a given site.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/federation/v1/sites/{site_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()